SLCORE-2061 SLCORE-2062 Remove on-demand omnisharp artifacts
